# contracts/openapi.yaml
openapi: 3.0.3
...
info:
  title: Clubify
  version: 1.0.0
  description: |
    API unique du backoffice, de l'application coach et du portail parent.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
paths:
...
  /auth/mfa/verify:
    post:
      tags: [auth]
      operationId: verifyMfa
      summary: Répond au défi de second facteur
      description: Accepte un code TOTP ou un code de secours à usage unique.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MfaVerifyRequest' }
      responses:
        '200':
          description: Jetons délivrés
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      summary: Renouvelle le jeton d'accès
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200':
          description: Nouveau couple de jetons
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /users:
    get:
      tags: [users]
      operationId: listUsers
      summary: Liste les utilisateurs du club
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/Role' }
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: Page d'utilisateurs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPage' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [users]
      operationId: createUser
      summary: Crée un utilisateur et son appartenance au club
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCreateRequest' }
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [users]
      operationId: getUser
      summary: Lit un utilisateur
      responses:
        '200':
          description: Utilisateur
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [users]
      operationId: updateUser
      summary: Modifie un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserUpdateRequest' }
      responses:
        '200':
          description: Utilisateur modifié
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{userId}/status:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [users]
      operationId: setUserStatus
      summary: Active ou désactive un utilisateur
      description: |
        La désactivation révoque les sessions. Le dernier administrateur d'un club
        ne peut pas être désactivé.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [active]
              properties:
                active: { type: boolean }
      responses:
        '204': { description: Statut modifié }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /users/{userId}/role:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [users]
      operationId: setUserRole
      summary: Attribue un rôle
      description: Le dernier administrateur d'un club ne peut pas être rétrogradé.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '204': { description: Rôle attribué }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /users/{userId}/permissions:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [users]
      operationId: getUserPermissions
      summary: Lit les permissions effectives et les surcharges
      responses:
        '200':
          description: Permissions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPermissions' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [users]
      operationId: setUserPermissionOverrides
      summary: Remplace les surcharges de permissions d'un utilisateur
      description: Réservé à l'administrateur du compte, non délégable.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: '#/components/schemas/PermissionOverride' }
      responses:
        '200':
          description: Permissions après modification
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserPermissions' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /users/{userId}/sessions:
    parameters:
      - $ref: '#/components/parameters/UserId'
    delete:
      tags: [users]
      operationId: revokeUserSessions
      summary: Ferme les sessions d'un utilisateur sans le désactiver
      responses:
        '204': { description: Sessions fermées }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{userId}/mfa:
    parameters:
      - $ref: '#/components/parameters/UserId'
    delete:
      tags: [users]
      operationId: resetUserMfa
      summary: Réinitialise le second facteur d'un utilisateur
      responses:
        '204': { description: Second facteur réinitialisé }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      description: >-
        Lignes par page. Ramenée dans les bornes si elle les dépasse : une page
        de mille lignes coûte autant au serveur qu'au navigateur, et personne ne
        la lit. Sans valeur, la taille retenue par le club s'applique.
      schema: { type: integer, minimum: 1, maximum: 100 }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Authentification requise ou refusée
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Forbidden:
      description: Permission manquante
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Introuvable, y compris lorsque la donnée appartient à un autre club
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Conflict:
      description: Conflit avec l'état courant
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    UnprocessableEntity:
      description: Règle métier non respectée
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
...
  schemas:
...
    ProblemDetail:
      type: object
      description: RFC 9457. Le champ `code` est stable ; l'interface le traduit.
      required: [status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          example: user.email.alreadyUsed
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              code: { type: string }
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
...
    Role:
      type: string
      description: |
        ACCOUNT_ADMIN est l'administrateur du compte (décision 0028), à ne pas
        confondre avec FRONT_DESK, l'accueil.
      enum: [ACCOUNT_ADMIN, MANAGER, FRONT_DESK, COACH, ACCOUNTANT, PARENT]

    # --- auth
...
    MfaVerifyRequest:
      type: object
      required: [mfaChallengeId, code]
      properties:
        mfaChallengeId: { type: string, format: uuid }
        code: { type: string }
        trustDevice: { type: boolean, default: false }
        deviceLabel: { type: string }
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: >-
            Facultatif pour un navigateur : le jeton voyage alors dans le cookie
            HttpOnly posé à la connexion. Les clients sans cookie le passent ici.
    TokenPair:
      type: object
      required: [accessToken, refreshToken, expiresInSeconds]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresInSeconds: { type: integer }
        deviceToken:
          type: string
          description: Délivré lorsque l'appareil vient d'être marqué de confiance.
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, active]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        phone: { type: string }
        language: { type: string }
        role: { $ref: '#/components/schemas/Role' }
        active: { type: boolean }
        mfaEnabled: { type: boolean }
        lastLoginAt: { type: string, format: date-time }
    UserPage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/User' }
        page: { $ref: '#/components/schemas/PageMeta' }
    UserCreateRequest:
      type: object
      required: [email, firstName, lastName, role]
      properties:
        email: { type: string, format: email }
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        phone: { type: string }
        language: { type: string, default: fr }
        role: { $ref: '#/components/schemas/Role' }
        password: { type: string, format: password }
    UserUpdateRequest:
      type: object
      required: [firstName, lastName]
      properties:
        firstName: { type: string, minLength: 1 }
        lastName: { type: string, minLength: 1 }
        phone: { type: string }
        language: { type: string }
...
    PermissionOverride:
      type: object
      required: [code, granted]
      properties:
        code: { type: string }
        granted:
          type: boolean
          description: Vrai pour accorder en plus du rôle, faux pour retirer.
        parameter: {}
    UserPermissions:
      type: object
      required: [role, effective, overrides]
      properties:
        role: { $ref: '#/components/schemas/Role' }
        effective:
          type: array
          items: { type: string }
        overrides:
          type: array
          items: { $ref: '#/components/schemas/PermissionOverride' }

    # --- audit
...